requirements.txt
torchvision
torch
omegaconf
hydra-core
huggingface_hub
transformers